Rename nedb.js to index.js
