+++ src/api/transaction.js
@@ -1,6 +1,6 @@
 const { Config, getContractMetadata } = require('../config');
 const Utils = require('../utils');
-const { db } = require('../db/nedb');
+const { db } = require('../db');
 
 const DEFAULT_GAS_COST = formatGasCost(Config.DEFAULT_GAS_LIMIT * Config.DEFAULT_GAS_PRICE);
