+++ src/index.js
@@ -3,7 +3,7 @@ const _ = require('lodash');
 const BodhiServer = require('./server');
 const ServerConfig = require('./config');
 const Constants = require('./constants');
-const { initDB, deleteBodhiData } = require('./db/nedb');
+const { initDB, deleteBodhiData } = require('./db');
 const Utils = require('./utils');
 const EmitterHelper = require('./utils/emitterHelper');
 const { getLogger } = require('./utils/logger');
