+++ src/schema/resolvers.js
@@ -12,7 +12,7 @@ const topicEvent = require('../api/topic_event');
 const centralizedOracle = require('../api/centralized_oracle');
 const decentralizedOracle = require('../api/decentralized_oracle');
 const { Config, getContractMetadata } = require('../config');
-const DBHelper = require('../db/nedb').DBHelper;
+const DBHelper = require('../db').DBHelper;
 const { txState, phase } = require('../constants');
 const { calculateSyncPercent, getAddressBalances } = require('../sync');
 const Utils = require('../utils');
