+++ src/sync/index.js
@@ -7,7 +7,7 @@ const moment = require('moment');
 const BigNumber = require('bignumber.js');
 const { getContractMetadata, isMainnet } = require('../config');
 const { BLOCK_0_TIMESTAMP, SATOSHI_CONVERSION } = require('../constants');
-const { db, DBHelper } = require('../db/nedb');
+const { db, DBHelper } = require('../db');
 const updateTxDB = require('./update_tx');
 
 const Topic = require('../models/topic');
