+++ src/sync/update_tx.js
@@ -8,7 +8,7 @@ const bodhiToken = require('../api/bodhi_token');
 const eventFactory = require('../api/event_factory');
 const centralizedOracle = require('../api/centralized_oracle');
 const decentralizedOracle = require('../api/decentralized_oracle');
-const DBHelper = require('../db/nedb').DBHelper;
+const DBHelper = require('../db').DBHelper;
 const { Config, getContractMetadata } = require('../config');
 const { txState } = require('../constants');
 const Utils = require('../utils');
